+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - **Breaking** Replaced `ImageDownscale` with `ImageDownscaleMode`.
 * Add `Img::original_color_type` metadata.
 
+* Fix image encoding with premultiplied alpha.
 * Fix BGRA image tagged as mask.
 * Add `Length::simplify` and automatically apply it to parsed expressions.
 * Add `zng::text::txt` standalone property. It sets the child node of the widget to `Text!`.